requirements.txt
Flask
azure-cognitiveservices-vision-computervision
python-dotenv
json